Move all cluster utils to cluster state management domain
