+++ __tests__/V4AzureClusterManagement.js
@@ -7,6 +7,7 @@ import { getConfiguration } from 'model/services/metadata/configuration';
 import nock from 'nock';
 import { StatusCodes } from 'shared/constants';
 import { OrganizationsRoutes } from 'shared/constants/routes';
+import { getNumberOfNodes } from 'stores/cluster/utils';
 import {
   API_ENDPOINT,
   appCatalogsResponse,
@@ -27,7 +28,6 @@ import {
   v4ClustersResponse,
 } from 'testUtils/mockHttpCalls';
 import { renderRouteWithStore } from 'testUtils/renderUtils';
-import { getNumberOfNodes } from 'utils/clusterUtils';
 
 describe('V4AzureClusterManagement', () => {
   const minNodesCount = 3;
