+++ src/components/Cluster/ClusterDetail/V4ClusterDetailTable.js
@@ -5,9 +5,9 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { CSSBreakpoints, Providers } from 'shared/constants';
 import { selectResourcesV4 } from 'stores/cluster/selectors';
+import { isClusterCreating } from 'stores/cluster/utils';
 import { FlexRowWithTwoBlocksOnEdges, mq } from 'styles';
 import Button from 'UI/Button';
-import { isClusterCreating } from 'utils/clusterUtils';
 
 import CredentialInfoRow from './CredentialInfoRow';
 import NodesRunning from './NodesRunning';
